Implementacion de metodos en serrvice

+++ src/main/java/com/example/restaurants/repository/ICategoria.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICategoria extends JpaRepository<categoria,Integer> {
+public interface ICategoria extends JpaRepository<categoria,Long> {
 
 }
