+++ src/main/java/com/example/restaurants/repository/IReceta.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IReceta extends JpaRepository<receta, Integer> {
+public interface IReceta extends JpaRepository<receta, Long> {
     receta findByProducto(Long idProducto);
     List<receta> findByInsumo(Long idinsumo);
 }
